+++ lib/packages/wikibase/data-model/src/Statement/StatementList.php
@@ -38,14 +38,15 @@ class StatementList implements IteratorAggregate, Comparable, Countable {
 	 * @throws InvalidArgumentException
 	 */
 	public function __construct( $statements = array() ) {
-		$this->addStatements( $statements );
-	}
-
-	private function addStatements( $statements ) {
 		$this->assertAreStatements( $statements );
 
-		foreach ( $statements as $statement ) {
-			$this->statements[] = $statement;
+		if ( is_array( $statements ) ) {
+			$this->statements = $statements;
+		}
+		else {
+			foreach ( $statements as $statement ) {
+				$this->statements[] = $statement;
+			}
 		}
 	}
 
@@ -126,6 +127,25 @@ class StatementList implements IteratorAggregate, Comparable, Countable {
 	}
 
 	/**
+	 * @since 2.3
+	 *
+	 * @param PropertyId $id
+	 *
+	 * @return StatementList
+	 */
+	public function getWithPropertyId( PropertyId $id ) {
+		$statements = array();
+
+		foreach ( $this->statements as $statement ) {
+			if ( $statement->getPropertyId()->equals( $id ) ) {
+				$statements[] = $statement;
+			}
+		}
+
+		return new self( $statements );
+	}
+
+	/**
 	 * Returns a list of all Snaks on this StatementList. This includes at least the main snaks of
 	 * Claims, the snaks from Claim qualifiers, and the snaks from Statement References.
 	 *
@@ -143,6 +163,21 @@ class StatementList implements IteratorAggregate, Comparable, Countable {
 			foreach( $statement->getAllSnaks() as $snak ) {
 				$snaks[] = $snak;
 			}
+		}
+
+		return $snaks;
+	}
+
+	/**
+	 * @since 2.3
+	 *
+	 * @return Snak[]
+	 */
+	public function getMainSnaks() {
+		$snaks = array();
+
+		foreach ( $this->statements as $statement ) {
+			$snaks[] = $statement->getMainSnak();
 		}
 
 		return $snaks;